[refactor] 헤더 및 레이아웃 수정
- 헤더
  - "두뇌 트레이닝" 텍스트로 적혀있던 부분을 로고 이미지로 수정
  - 로고 이미지 클릭 시 관리자는 관리자 페이지로, 나머지는 메인 페이지로 이동
- 사이드바
  - 불필요한 사이드바 컴포넌트 삭제 및 관리자 사이드바 위치 수정
